renaming auth utils to login-handlers, more descriptive
